backend/requirements.txt
# FastAPI & Server
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# JSON parsing – orjson is non-recursive (no recursion limit issues) and 2-3x faster
orjson>=3.11.0

# Database
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.35

# LLM – pick your provider(s)
openai==1.51.0
httpx<0.28.0
google-generativeai>=0.8.0
groq>=0.11.0

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.2
markdown==3.7

# Config
pydantic-settings==2.5.0

# SSE
sse-starlette==2.1.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
